Testing my viewBinding setup with the viewmodel. It is not working!
I made a setup where button press should update the textview as a first step. This is to check that my 'basic' viewModel and observer pattern implementation is working. On button press the textview did not get updated so I placed it outside the observer, just to check and it is not updating there as well! Need to investigate further.

I have left the dataclass alone for now to be implemented later once the update of variables in the viewModel will be working fine.